+++ include/lwip/apps/snmp_opts.h
@@ -39,8 +39,6 @@
 
 #include "lwip/opt.h"
 
-#include <arpa/inet.h>
-
 /**
  * @defgroup snmp_opts Options
  * @ingroup snmp