+++ include/lwip/def.h
@@ -82,29 +82,16 @@ extern "C" {
 #endif
 #endif
 
-#if BYTE_ORDER == BIG_ENDIAN
-#define lwip_htons(x) ((u16_t)(x))
-#define lwip_ntohs(x) ((u16_t)(x))
-#define lwip_htonl(x) ((u32_t)(x))
-#define lwip_ntohl(x) ((u32_t)(x))
-#define PP_HTONS(x)   ((u16_t)(x))
-#define PP_NTOHS(x)   ((u16_t)(x))
-#define PP_HTONL(x)   ((u32_t)(x))
-#define PP_NTOHL(x)   ((u32_t)(x))
-#else /* BYTE_ORDER != BIG_ENDIAN */
-#ifndef lwip_htons
-u16_t lwip_htons(u16_t x);
-#endif
-#define lwip_ntohs(x) lwip_htons(x)
+#include <zephyr/net/net_ip.h>
+#include <zephyr/sys/byteorder.h>
 
-#ifndef lwip_htonl
-u32_t lwip_htonl(u32_t x);
-#endif
-#define lwip_ntohl(x) lwip_htonl(x)
+#define lwip_htons(x) net_htons(x)
+#define lwip_ntohs(x) net_ntohs(x)
+#define lwip_htonl(x) net_htonl(x)
+#define lwip_ntohl(x) net_ntohl(x)
 
-/* These macros should be calculated by the preprocessor and are used
-   with compile-time constants only (so that there is no little-endian
-   overhead at runtime). */
+/* These macros are calculated by the preprocessor and are used with
+   compile-time constants only, so that there is no runtime cost. */
 #define PP_HTONS(x) ((u16_t)((((x) & (u16_t)0x00ffU) << 8) | (((x) & (u16_t)0xff00U) >> 8)))
 #define PP_NTOHS(x) PP_HTONS(x)
 #define PP_HTONL(x) ((((x) & (u32_t)0x000000ffUL) << 24) | \
@@ -112,7 +99,6 @@ u32_t lwip_htonl(u32_t x);
                      (((x) & (u32_t)0x00ff0000UL) >>  8) | \
                      (((x) & (u32_t)0xff000000UL) >> 24))
 #define PP_NTOHL(x) PP_HTONL(x)
-#endif /* BYTE_ORDER == BIG_ENDIAN */
 
 /* Provide usual function names as macros for users, but this can be turned off */
 #ifndef LWIP_DONT_PROVIDE_BYTEORDER_FUNCTIONS